requirements.txt
psycopg2-binary
streamlit
openai
pgvector
sqlalchemy
